feat: anamnese ajusatada
